+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     + `+[NSDate dateFromHTTPDateString:]`
     + `NSDateFormatter (ESAdditions)` category
     + NSArray methods: `objectOrNilAtIndex:`, `-previousObjectToIndex:`, `-previousObjectToObject:`, `-nextObjectToIndex:`, `-nextObjectToObject:`
+    + NSMutableArray methods: `-removeFirstObject`, `-shiftFirstObject`, `-popLastObject`
     + NSOrderedSet methods: `objectOrNilAtIndex:`
     + `-[NSUserDefaults setObject:forKeyPath:]`
     + `-[UIDevice modelName]`